Feat (ED-77): refactor sheet to json script
