+++ scripts/utils/credentials_sheet_handler.py
@@ -2,10 +2,6 @@ import os
 import json
 from dotenv import load_dotenv
 
-from google.auth.transport.requests import Request
-from google.oauth2.credentials import Credentials
-from google_auth_oauthlib.flow import InstalledAppFlow
-
 
 def check_environment_variables(client_id_denes_sheet, client_secret_denes_sheet):
     if not client_id_denes_sheet or not client_secret_denes_sheet:
@@ -49,21 +45,6 @@ def main():
     client_secret_denes_sheet = os.getenv('CLIENT_SECRET_DENES_SHEET')
     credentials_path = get_absolute_path(credentials_relative_path)
     update_credentials(client_id_denes_sheet, client_secret_denes_sheet, credentials_path)
-
-
-def get_credentials(credentials_path, token_filename, scopes):
-    credentials = None
-    if os.path.exists(get_absolute_path(token_filename)):
-        credentials = Credentials.from_authorized_user_file(get_absolute_path(token_filename), scopes)
-    if not credentials or not credentials.valid:
-        if credentials and credentials.expired and credentials.refresh_token:
-            credentials.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(credentials_path, scopes)
-            credentials = flow.run_local_server(port=0)
-        with open(get_absolute_path(token_filename), "w") as token:
-            token.write(credentials.to_json())
-    return credentials
 
 
 if __name__ == "__main__":
